backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "google-generativeai>=0.8.6",
    "python-dotenv>=1.2.2",
    "supabase>=2.28.0",
    "uvicorn>=0.41.0",
]
